檢查的函式攏囥做伙～～
