+++ gi2_liau7_khoo3/settings/base.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     '語言模型',
     '口語辭典',
     '本調辭典',
+    '檢查工具',
 ]
 
 MIDDLEWARE = [
