+++ 檢查工具/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class 檢查工具Config(AppConfig):
+    name = '檢查工具'
