+++ 語料庫/models.py
@@ -104,13 +104,3 @@ class 語料狀況表(models.Model):
 
     def __str__(self):
         return '{} {}'.format(self.pk, self.狀況.split('：')[-1])
-
-
-class 對齊狀態表(models.Model):
-    語料 = models.OneToOneField(
-        '語料表', default=None, related_name='對齊狀態',
-        on_delete=models.CASCADE)
-    狀態 = models.CharField(max_length=30)
-
-    def __str__(self):
-        return self.狀態
